rework notifications
